Work till 5 pm 12th June

+++ PreferenceToLongerPaths/src/com/iitb/gise/Graph.java
@@ -48,7 +48,7 @@ public class Graph {
 			if(n.getNodeType() == Node.SOURCE)
 				n.setTravelTime(0);
 			else
-				n.setTravelTime(Double.MAX_VALUE);
+				n.setTravelTime(Integer.MAX_VALUE);
 		}
 	}
 }
